Refactor debug routes into a separate file
