+++ server/views/pages/debug/incidentDetails.njk
@@ -1,48 +1,51 @@
-{% extends "../partials/layout.njk" %}
+{% extends "../../partials/layout.njk" %}
+
 {% from "govuk/components/table/macro.njk" import govukTable %}
+
 {% set pageTitle = applicationName + " - Incident" %}
 {% set mainClasses = "govuk-body" %}
+
 {% block content %}
   <h1>Incident '{{ event.title }}'</h1>
 
   {{ govukTable({
-  caption: "Incident details",
-  captionClasses: "govuk-table__caption--m",
-  firstCellIsHeader: true,
-  rows: [
-    [
-      {
-        text: "Title"
-      },
-      {
-        text: event.title
-      }
-    ],
-    [
-      {
-        text: "Reference"
-      },
-      {
-        text: event.eventReference
-      }
-    ],
-    [
-      {
-        text: "When occurred"
-      },
-      {
-        text: event.eventDateAndTime | dateAndTime
-      }
-    ],
-    [
-      {
-        text: "Description"
-      },
-      {
-        text: event.description
-      }
+    caption: "Incident details",
+    captionClasses: "govuk-table__caption--m",
+    firstCellIsHeader: true,
+    rows: [
+      [
+        {
+          text: "Title"
+        },
+        {
+          text: event.title
+        }
+      ],
+      [
+        {
+          text: "Reference"
+        },
+        {
+          text: event.eventReference
+        }
+      ],
+      [
+        {
+          text: "When occurred"
+        },
+        {
+          text: event.eventDateAndTime | dateAndTime
+        }
+      ],
+      [
+        {
+          text: "Description"
+        },
+        {
+          text: event.description
+        }
+      ]
     ]
-  ]
   }) }}
 
   {%- for report in event.reports -%}
