--- server/views/pages/events.njk
@@ -1,11 +1,0 @@
-{% extends "../partials/layout.njk" %}
-{% set pageTitle = applicationName + " - Home" %}
-{% set mainClasses = "app-container govuk-body" %}
-{% block content %}
-  <h1>Incident events</h1>
-  <p>API Response for <pre>GET /incident-events/</pre>
-    endpoint</p>
-  <pre>
-    {{ response | dump(2) }}
-  </pre>
-{% endblock %}
